Refactor drag code
